FIX - Refactoring folder structure and flow
